[level 1] Title: 내적, Time: 0.04 ms, Memory: 75.9 MB -BaekjoonHub

+++ 프로그래머스/1/70128. 내적/내적.java
@@ -1,9 +1,9 @@
 class Solution {
     public int solution(int[] a, int[] b) {
-        long answer = 0;
+        int answer = 0;
         for(int i=0; i<a.length; i++){
             answer += a[i] * b[i];
         }
-        return (int)answer;
+        return answer;
     }
 }
